rng: fix mwc64.inRange + add test

+++ src/rng/mwc.h
@@ -73,7 +73,7 @@ public:
   }
 #else
   //  using xoro::xoro;
-  MWC(uint32_t s1, uint32_t s2)
+  MWC(uint64_t s1, uint64_t s2)
   // : xoro::xoro(s1, s2)
   : MWC64::MWC64(s1, s2) {
     //    std::cout << sizeof(RealMWC) << std::endl;
@@ -86,7 +86,7 @@ public:
 
     //    return min + next() % range;
     // adapted from https://lemire.me/blog/2016/06/27/a-fast-alternative-to-the-modulo-reduction/
-    return min + (((uint64_t)next() * (uint64_t)range) >> 32);
+    return min + (((uint64_t)(uint32_t)next() * (uint64_t)range) >> 32);
   }
 };
 #endif
